+++ libs-scala/security-evidence/tag/src/scala/com/daml/security/evidence/tag/MissingTest.scala
@@ -1,0 +1,11 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.security.evidence.tag
+
+/** A tag for a missing test case.
+  *
+  * Similar to a to-do entry for test coverage but can be included in the inventory.
+  * Alternatively one can use `ignore` with the appropriate tag already applied.
+  */
+final case class MissingTest(missingCoverage: String) extends EvidenceTag
