+++ libs-scala/security-evidence/tag/src/scala/com/daml/security/evidence/tag/Operability.scala
@@ -1,0 +1,20 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.security.evidence.tag
+
+object Operability {
+
+  trait OperabilityTestSuite
+
+  object OperabilityTestSuite {}
+
+  final case class OperabilityTest(
+      component: String,
+      dependency: String,
+      setting: String,
+      cause: String,
+      remediation: String,
+  ) extends EvidenceTag
+
+}
